initial models and sql
took the time to review the code and understand which jobs i should modify. as of now it's going to be mostly focused on const_cf_pslq, with secondary priority on families and day_summary, and tertiary priority on cf_to_cf_pslq. Should probably also add there some tool for converting "single constant connections" to multi_constant_connections